+++ internal/pkg/clients/clients_test.go
@@ -35,7 +35,7 @@ func TestSupportedClients(t *testing.T) {
 		{"execution", "mainnet", AllClients["execution"], false},
 		{"consensus", "mainnet", AllClients["consensus"], false},
 		{"validator", "mainnet", AllClients["validator"], false},
-		{"distributedValidator", "goerli", AllClients["distributedValidator"], false},
+		{"distributedValidator", "holesky", AllClients["distributedValidator"], false},
 		{"random", "mainnet", []string{}, true},
 	}
 
@@ -147,7 +147,7 @@ func TestClients(t *testing.T) {
 				"distributedValidator": {"charon"},
 			},
 			[]string{"consensus", "execution", "validator", "distributedValidator"},
-			"goerli",
+			"holesky",
 			false,
 		},
 	}