+++ internal/pkg/generate/generate_scripts.go
@@ -20,6 +20,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"slices"
 	"strings"
 	"text/template"
 
@@ -266,6 +267,7 @@ func ComposeFile(gd *GenData, at io.Writer) error {
 			gd.MevBoostEndpoint = fmt.Sprintf("%s:%v", configs.DefaultMevBoostEndpoint, gd.Ports["MevPort"])
 		}
 	}
+	gd.MevBoostService = slices.Contains(gd.Services, "mev-boost")
 
 	if gd.Distributed {
 		// Check for distributed validator